pyproject.toml
[tool.poetry]
authors = ["Promptworks <hello@promptworks.com>"]
description = ""
license= "MIT"
name = "atat"
version = "0.1.0"
repository = "https://github.com/dod-ccpo/atst"
homepage = "https://github.com/dod-ccpo/atst"



[tool.poetry.dependencies]
python = "3.7.3"
alembic = "*"
azure-storage-blob = "*"
cachetools = "^4.1.1"
celery = "*"
flask = "*"
flask-assets = "*"
flask-session = "*"
flask-sqlalchemy = "*"
flask-wtf = "*"
lockfile = "*"
msrestazure = "*"
pendulum = "*"
"psycopg2-binary" = "*"
packaging = "*"
pydantic = "*"
pyopenssl = "*"
PyYAML = "*"
redis = "*"
requests = "*"
sqlalchemy = ">=1.3.12"
sqlalchemy-json = "*"
transitions = "*"
Unipath = "*"
webassets = "*"
werkzeug = "*"
email-validator = '*'
python3-saml = "^1.9.0"
kubernetes = "^11.0.0"
ansible = "^2.9.11"
openshift = "^0.11.2"

[tool.poetry.dev-dependencies]
bandit = "*"
pytest = "*"
ipython = "*"
ipdb = "*"
pylint = "*"
factory-boy = "*"
pytest-flask = "*"
pytest-env = "*"
pytest-cov = "*"
selenium = "*"
honcho = "*"
blinker = "*"
pytest-mock = "*"
detect-secrets = "*"
beautifulsoup4 = "*"
mypy = "*"
rope = "*"
black = "^19.10b0"
pytest-depends = "^1.0.1"


[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[tool.isort]
profile = 'black'
